vecindex: extract vecsettings to break dependency for scbuildstmt
Release note: None
