+++ pkg/BUILD.bazel
@@ -2472,6 +2472,7 @@ GO_TARGETS = [
     "//pkg/sql/vecindex/vecencoding:vecencoding_test",
     "//pkg/sql/vecindex/vecpb:vecpb",
     "//pkg/sql/vecindex/vecpb:vecpb_test",
+    "//pkg/sql/vecindex/vecsettings:vecsettings",
     "//pkg/sql/vecindex/vecstore/vecstorepb:vecstorepb",
     "//pkg/sql/vecindex/vecstore:vecstore",
     "//pkg/sql/vecindex/vecstore:vecstore_test",
