+++ pkg/sql/BUILD.bazel
@@ -551,6 +551,7 @@ go_library(
         "//pkg/sql/unsafesql",
         "//pkg/sql/vecindex",
         "//pkg/sql/vecindex/vecpb",
+        "//pkg/sql/vecindex/vecsettings",
         "//pkg/sql/vecindex/vecstore",
         "//pkg/sql/vtable",
         "//pkg/storage",
